Cambios en las clases para reflejar el nuevo sistema de ventas
