--- src/main/java/negocio/Foto.java
@@ -1,4 +1,0 @@
-package negocio;
-
-public class Foto {
-}
