--- src/main/java/negocio/SituacionLaboral.java
@@ -1,7 +1,0 @@
-package negocio;
-
-public enum SituacionLaboral {
-    DEPENDIENTE,
-    INDEPEDIENTE,
-    DESEMPLEADO
-}
