--- src/main/java/negocio/TipoDeProducto.java
@@ -1,12 +1,0 @@
-package negocio;
-
-public enum TipoDeProducto {
-    MOUSE,
-    TECLADO,
-    GABINETE,
-    MOTHER,
-    RAM,
-    COOLER,
-    PLACA_VIDEO,
-    FUENTE
-}
