+++ src/main/java/negocio/componente/Foto.java
@@ -1,0 +1,4 @@
+package negocio.componente;
+
+public class Foto {
+}
