+++ src/main/java/negocio/envio/Correo.java
@@ -1,0 +1,8 @@
+package negocio.envio;
+
+public class Correo implements TipoDeEnvio{
+
+    public Integer calcularCosto() {
+        return 0;
+    }
+}
