+++ src/main/java/negocio/envio/Moto.java
@@ -1,0 +1,8 @@
+package negocio.envio;
+
+public class Moto implements TipoDeEnvio{
+
+    public Integer calcularCosto() {
+        return 0;
+    }
+}
