+++ src/main/java/negocio/envio/Rapido.java
@@ -1,0 +1,8 @@
+package negocio.envio;
+
+public class Rapido implements TipoDeEnvio{
+
+    public Integer calcularCosto() {
+        return 0;
+    }
+}
