+++ src/main/java/negocio/envio/TipoDeEnvio.java
@@ -1,0 +1,5 @@
+package negocio.envio;
+
+public interface TipoDeEnvio {
+    Integer calcularCosto();
+}
